feat: delete unused call of interceptor headers function electron | [#194]
